+++ ti99/cpu/tms9900/tms9901.c
@@ -407,6 +407,7 @@ ITCM_CODE u16 TMS9901_ReadCRU(u16 cruAddress, u8 num)
                             // This handles both Keybaord and Joystick (P1 and P2) inputs in a unified manner... to the TI-99/4a, it's all the same.
                             // ------------------------------------------------------------------------------------------------------------------------
                             u8 column = (tms9901.PinState[PIN_COL3]<<2) | (tms9901.PinState[PIN_COL2]<<1) | (tms9901.PinState[PIN_COL1]<<0);
+                            tms9901.KeyColsScanned |= (1 << column);    // note that this column was looked at
                             if (tms9901.Keyboard[TIKeys[cruA-3][column]]) bitState = 0;
                         }
                         break;